sync

+++ Assets/Scripts/SpawnCoins.cs
@@ -20,7 +20,7 @@ public class SpawnCoins : MonoBehaviour
     {
         for (int i = 0; i < coinSpawns.Length; i++)
         {
-            int coinFlip = Random.Range(0, 2);
+            float coinFlip = Random.Range(0f, 2f);
             if (coinFlip>0)
             {
                 Instantiate(coin, coinSpawns[i].position, Quaternion.identity);
